refactor: SKillsSection
